extract deep_merge into src/utils
